requirements.txt
streamlit
pandas
numpy
python-dateutil
langchain-groq
sqlite3